+++ project.go
@@ -7,6 +7,8 @@ import (
 	"context"
 	"fmt"
 	"net/url"
+
+	"github.com/hashicorp/jsonapi"
 )
 
 // Compile-time proof of interface implementation.
@@ -63,6 +65,8 @@ type Project struct {
 
 	Description string `jsonapi:"attr,description"`
 
+	AutoDestroyActivityDuration jsonapi.NullableAttr[string] `jsonapi:"attr,auto-destroy-activity-duration,omitempty"`
+
 	// Relations
 	Organization *Organization `jsonapi:"relation,organization"`
 }
@@ -100,6 +104,11 @@ type ProjectCreateOptions struct {
 
 	// Associated TagBindings of the project.
 	TagBindings []*TagBinding `jsonapi:"relation,tag-bindings,omitempty"`
+
+	// Optional: For all workspaces in the project, the period of time to wait 
+	// after workspace activity to trigger a destroy run. The format should roughly 
+	// match a Go duration string limited to days and hours, e.g. "24h" or "1d".
+	AutoDestroyActivityDuration jsonapi.NullableAttr[string] `jsonapi:"attr,auto-destroy-activity-duration,omitempty"`
 }
 
 // ProjectUpdateOptions represents the options for updating a project
@@ -119,6 +128,11 @@ type ProjectUpdateOptions struct {
 	// Associated TagBindings of the project. Note that this will replace
 	// all existing tag bindings.
 	TagBindings []*TagBinding `jsonapi:"relation,tag-bindings,omitempty"`
+
+	// Optional: For all workspaces in the project, the period of time to wait 
+	// after workspace activity to trigger a destroy run. The format should roughly 
+	// match a Go duration string limited to days and hours, e.g. "24h" or "1d".
+	AutoDestroyActivityDuration jsonapi.NullableAttr[string] `jsonapi:"attr,auto-destroy-activity-duration,omitempty"`
 }
 
 // ProjectAddTagBindingsOptions represents the options for adding tag bindings